rename datasets to avoid hf conflict
