+++ llmfoundry/data/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2022 MosaicML LLM Foundry authors
 # SPDX-License-Identifier: Apache-2.0
 
-from llmfoundry.data.datasets import ConcatTokensDataset, NoConcatDataset
+from llmfoundry.data.data import ConcatTokensDataset, NoConcatDataset
 from llmfoundry.data.denoising import (MixtureOfDenoisersCollator,
                                        build_text_denoising_dataloader)
 from llmfoundry.data.finetuning import (Seq2SeqFinetuningCollator,
