+++ scripts/data_prep/convert_dataset_hf.py
@@ -10,13 +10,12 @@ from enum import Enum
 from typing import Dict, Iterable, Optional, Union
 
 import datasets as hf_datasets
-import numpy as np
 from streaming import MDSWriter
 from torch.utils.data import DataLoader, IterableDataset
 from tqdm import tqdm
 from transformers import AutoTokenizer, PreTrainedTokenizerBase
 
-from llmfoundry.data.datasets import ConcatTokensDataset, NoConcatDataset
+from llmfoundry.data import ConcatTokensDataset, NoConcatDataset
 
 
 class ConcatMode(Enum):
